Refactor some code out of main conditions component
