+++ src/fhir/conditions.ts
@@ -9,6 +9,8 @@ import {
   searchLensRecords,
 } from "./search-helpers";
 import {
+  SYSTEM_CONDITION_CLINICAL,
+  SYSTEM_CONDITION_VERIFICATION_STATUS,
   SYSTEM_ICD10,
   SYSTEM_ICD10_CM,
   SYSTEM_ICD9,
@@ -48,6 +50,38 @@ export type QueryKeyPatientConditions = [
   ConditionFilters
 ];
 export type QueryKeyOtherProviderConditions = [string, string | undefined];
+
+export function getNewCondition(patientId: string) {
+  const newCondition: fhir4.Condition = {
+    resourceType: "Condition",
+    subject: {
+      type: "Patient",
+      reference: `Patient/${patientId}`,
+    },
+    clinicalStatus: {
+      coding: [
+        {
+          system: SYSTEM_CONDITION_CLINICAL,
+          code: "active",
+          display: "Active",
+        },
+      ],
+      text: "active",
+    },
+    verificationStatus: {
+      coding: [
+        {
+          system: SYSTEM_CONDITION_VERIFICATION_STATUS,
+          code: "confirmed",
+          display: "Confirmed",
+        },
+      ],
+      text: "confirmed",
+    },
+  };
+
+  return newCondition;
+}
 
 export async function getPatientConditions(
   queryParams: QueryFunctionContext<QueryKeyPatientConditions>
